generated_projects/project_w1-5/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
Jinja2==3.1.2